requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Web interface
streamlit>=1.25.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
click>=8.1.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for enhanced functionality
accelerate>=0.20.0
bitsandbytes>=0.41.0
optimum>=1.12.0

# Jupyter notebook support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0

# Additional ML libraries
sentence-transformers>=2.2.0
textblob>=0.17.0
vaderSentiment>=3.3.0
